+++ mkt/reviewers/tests/test_views.py
@@ -51,8 +51,8 @@ from mkt.tags.models import Tag
 from mkt.versions.models import Version
 from mkt.webapps.models import Webapp
 from mkt.webapps.tests.test_models import PackagedFilesMixin
+from mkt.zadmin.models import get_config, set_config
 from users.models import UserProfile
-from zadmin.models import get_config, set_config
 
 
 TEST_PATH = path.dirname(path.abspath(__file__))
