+++ mkt/zadmin/templates/zadmin/manifest.html
