+++ mkt/zadmin/templates/zadmin/update-prices.html
