+++ mkt/zadmin/tests/test_models.py
@@ -1,7 +1,7 @@
 from nose.tools import eq_
 
 from amo.tests import TestCase
-from zadmin.models import DownloadSource
+from mkt.zadmin.models import DownloadSource
 
 
 class TestDownloadSource(TestCase):
