+++ mkt/zadmin/tests/test_views.py
@@ -1,18 +1,259 @@
+# -*- coding: utf-8 -*-
+import csv
+from cStringIO import StringIO
+
 from django.conf import settings
-from django.core.urlresolvers import reverse
+from django.core import mail
+from django.core.cache import cache
 
 from nose.tools import eq_
-from pyquery import PyQuery as pq
 
 import amo
 import amo.tests
+from addons.models import Addon
+from amo.urlresolvers import reverse
+from mkt.access.models import Group, GroupUser
 from mkt.reviewers.models import RereviewQueue
 from mkt.site.fixtures import fixture
 from users.models import UserProfile
 
+from ..forms import DevMailerForm
+from ..models import EmailPreviewTopic
+
+
+class TestEmailPreview(amo.tests.TestCase):
+    fixtures = fixture('user_admin', 'group_admin', 'user_admin_group',
+                       'webapp_337141')
+
+    def setUp(self):
+        assert self.client.login(username='admin@mozilla.com',
+                                 password='password')
+        addon = Addon.objects.get(pk=337141)
+        self.topic = EmailPreviewTopic(addon)
+
+    def test_csv(self):
+        self.topic.send_mail('the subject', u'Hello Ivan Krsti\u0107',
+                             from_email='admin@mozilla.org',
+                             recipient_list=['funnyguy@mozilla.org'])
+        r = self.client.get(reverse('zadmin.email_preview_csv',
+                            args=[self.topic.topic]))
+        eq_(r.status_code, 200)
+        rdr = csv.reader(StringIO(r.content))
+        eq_(rdr.next(), ['from_email', 'recipient_list', 'subject', 'body'])
+        eq_(rdr.next(), ['admin@mozilla.org', 'funnyguy@mozilla.org',
+                         'the subject', 'Hello Ivan Krsti\xc4\x87'])
+
+
+class TestMemcache(amo.tests.TestCase):
+    fixtures = fixture('user_admin', 'group_admin', 'user_admin_group')
+
+    def setUp(self):
+        self.url = reverse('zadmin.memcache')
+        cache.set('foo', 'bar')
+        self.client.login(username='admin@mozilla.com', password='password')
+
+    def test_login(self):
+        self.client.logout()
+        eq_(self.client.get(self.url).status_code, 302)
+
+    def test_can_clear(self):
+        self.client.post(self.url, {'yes': 'True'})
+        eq_(cache.get('foo'), None)
+
+    def test_cant_clear(self):
+        self.client.post(self.url, {'yes': 'False'})
+        eq_(cache.get('foo'), 'bar')
+
+
+class TestElastic(amo.tests.ESTestCase):
+    fixtures = fixture('user_admin', 'group_admin', 'user_admin_group')
+
+    def setUp(self):
+        self.url = reverse('zadmin.elastic')
+        self.client.login(username='admin@mozilla.com', password='password')
+
+    def test_login(self):
+        self.client.logout()
+        self.assertRedirects(self.client.get(self.url),
+            reverse('users.login') + '?to=/admin/elastic')
+
+
+class TestEmailDevs(amo.tests.TestCase):
+    fixtures = fixture('user_admin', 'group_admin', 'user_admin_group',
+                       'webapp_337141')
+
+    def setUp(self):
+        self.login('admin')
+        self.addon = Addon.objects.get(pk=337141)
+
+    def post(self, recipients=None, subject='subject', message='msg',
+             preview_only=False):
+        return self.client.post(reverse('zadmin.email_devs'),
+                                dict(recipients=recipients, subject=subject,
+                                     message=message,
+                                     preview_only=preview_only))
+
+    def test_preview(self):
+        self.addon.update(premium_type=amo.ADDON_PREMIUM)
+        res = self.post(recipients='payments', preview_only=True)
+        self.assertNoFormErrors(res)
+        preview = EmailPreviewTopic(topic='email-devs')
+        eq_([e.recipient_list for e in preview.filter()],
+            ['steamcube@mozilla.com'])
+        eq_(len(mail.outbox), 0)
+
+    def test_only_apps_with_payments(self):
+        self.addon.update(premium_type=amo.ADDON_PREMIUM)
+        res = self.post(recipients='payments')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 1)
+
+        mail.outbox = []
+        self.addon.update(status=amo.STATUS_PENDING)
+        res = self.post(recipients='payments')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 1)
+
+        mail.outbox = []
+        self.addon.update(status=amo.STATUS_DELETED)
+        res = self.post(recipients='payments')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 0)
+
+    def test_only_free_apps_with_new_regions(self):
+        res = self.post(recipients='free_apps_region_enabled')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 0)
+        mail.outbox = []
+        res = self.post(recipients='free_apps_region_disabled')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 1)
+
+        mail.outbox = []
+        self.addon.update(enable_new_regions=True)
+        res = self.post(recipients='free_apps_region_enabled')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 1)
+        mail.outbox = []
+        res = self.post(recipients='free_apps_region_disabled')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 0)
+
+    def test_only_apps_with_payments_and_new_regions(self):
+        self.addon.update(premium_type=amo.ADDON_PREMIUM)
+        res = self.post(recipients='payments_region_enabled')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 0)
+        mail.outbox = []
+        res = self.post(recipients='payments_region_disabled')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 1)
+
+        mail.outbox = []
+        self.addon.update(enable_new_regions=True)
+        res = self.post(recipients='payments_region_enabled')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 1)
+        mail.outbox = []
+        res = self.post(recipients='payments_region_disabled')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 0)
+
+    def test_only_desktop_apps(self):
+        from addons.models import AddonDeviceType
+        AddonDeviceType.objects.create(addon=self.addon,
+            device_type=amo.DEVICE_MOBILE.id)
+        res = self.post(recipients='desktop_apps')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 0)
+
+        mail.outbox = []
+        AddonDeviceType.objects.create(addon=self.addon,
+            device_type=amo.DEVICE_DESKTOP.id)
+        res = self.post(recipients='desktop_apps')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 1)
+
+        mail.outbox = []
+        self.addon.update(status=amo.STATUS_PENDING)
+        res = self.post(recipients='desktop_apps')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 1)
+
+        mail.outbox = []
+        self.addon.update(status=amo.STATUS_DELETED)
+        res = self.post(recipients='desktop_apps')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 0)
+
+    def test_only_apps(self):
+        res = self.post(recipients='apps')
+        self.assertNoFormErrors(res)
+        eq_(len(mail.outbox), 1)
+
+    def test_ignore_deleted_always(self):
+        self.addon.update(status=amo.STATUS_DELETED)
+        for name, label in DevMailerForm._choices:
+            res = self.post(recipients=name)
+            self.assertNoFormErrors(res)
+            eq_(len(mail.outbox), 0)
+
+    def test_exclude_pending_for_addons(self):
+        self.addon.update(status=amo.STATUS_PENDING)
+        for name, label in DevMailerForm._choices:
+            if name in ('payments', 'desktop_apps'):
+                continue
+            res = self.post(recipients=name)
+            self.assertNoFormErrors(res)
+            eq_(len(mail.outbox), 0)
+
+
+class TestPerms(amo.tests.TestCase):
+    fixtures = fixture('user_admin', 'group_admin', 'user_admin_group',
+                       'user_999')
+
+    def test_admin_user(self):
+        # Admin should see views with Django's perm decorator and our own.
+        assert self.client.login(username='admin@mozilla.com',
+                                 password='password')
+        eq_(self.client.get(reverse('zadmin.index')).status_code, 200)
+        eq_(self.client.get(reverse('zadmin.settings')).status_code, 200)
+
+    def test_staff_user(self):
+        # Staff users have some privileges.
+        user = UserProfile.objects.get(email='regular@mozilla.com')
+        group = Group.objects.create(name='Staff', rules='AdminTools:View')
+        GroupUser.objects.create(group=group, user=user)
+        assert self.client.login(username='regular@mozilla.com',
+                                 password='password')
+        eq_(self.client.get(reverse('zadmin.index')).status_code, 200)
+        eq_(self.client.get(reverse('zadmin.settings')).status_code, 200)
+
+    def test_sr_reviewers_user(self):
+        # Sr Reviewers users have only a few privileges.
+        user = UserProfile.objects.get(email='regular@mozilla.com')
+        group = Group.objects.create(name='Sr Reviewer',
+                                     rules='ReviewerAdminTools:View')
+        GroupUser.objects.create(group=group, user=user)
+        assert self.client.login(username='regular@mozilla.com',
+                                 password='password')
+        eq_(self.client.get(reverse('zadmin.index')).status_code, 200)
+        eq_(self.client.get(reverse('zadmin.settings')).status_code, 403)
+
+    def test_unprivileged_user(self):
+        # Unprivileged user.
+        assert self.client.login(username='regular@mozilla.com',
+                                 password='password')
+        eq_(self.client.get(reverse('zadmin.index')).status_code, 403)
+        eq_(self.client.get(reverse('zadmin.settings')).status_code, 403)
+        # Anonymous users should also get a 403.
+        self.client.logout()
+        self.assertRedirects(self.client.get(reverse('zadmin.index')),
+                             reverse('users.login') + '?to=/admin/')
+
 
 class TestHome(amo.tests.TestCase):
-    fixtures = ['base/users']
+    fixtures = fixture('user_admin', 'group_admin', 'user_admin_group')
 
     def setUp(self):
         self.client.login(username='admin@mozilla.com', password='password')
@@ -25,7 +266,7 @@ class TestHome(amo.tests.TestCase):
 
 
 class TestGenerateError(amo.tests.TestCase):
-    fixtures = ['base/users']
+    fixtures = fixture('user_admin', 'group_admin', 'user_admin_group')
 
     def setUp(self):
         self.client.login(username='admin@mozilla.com', password='password')
@@ -101,24 +342,9 @@ class TestGenerateError(amo.tests.TestCase):
         eq_(msg.type, 'sentry')
 
 
-class TestAddonAdmin(amo.tests.TestCase):
-    fixtures = ['base/users', 'base/337141-steamcube', 'base/addon_3615']
-
-    def setUp(self):
-        self.login('admin@mozilla.com')
-        self.url = reverse('admin:addons_addon_changelist')
-
-    def test_no_webapps(self):
-        res = self.client.get(self.url, follow=True)
-        eq_(res.status_code, 200)
-        doc = pq(res.content)
-        rows = doc('#result_list tbody tr')
-        eq_(rows.length, 1)
-        eq_(rows.find('a').attr('href'), '/admin/models/addons/addon/337141/')
-
-
 class TestManifestRevalidation(amo.tests.TestCase):
-    fixtures = fixture('webapp_337141') + ['base/users']
+    fixtures = fixture('user_admin', 'group_admin', 'user_admin_group',
+                       'webapp_337141', 'user_999')
 
     def setUp(self):
         self.url = reverse('zadmin.manifest_revalidation')
